Move ResponseStatus to ResponseUtil
